LayoutTests/fast/frames/seamless/seamless-css-cascade.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<style>
iframe { color: orange; }
div { color: white; }
#one { color: red; }
#two { color: blue; }
</style>
<iframe id="iframe" webkitseamless src="resources/css-cascade-child.html"></iframe>
<script>
window.onload = function () {
    window.iframe = document.getElementById("iframe");
    window.one = iframe.contentDocument.getElementById("one");
    window.two = iframe.contentDocument.getElementById("two");
    window.three = iframe.contentDocument.getElementById("three");

    // Spec: In a CSS-supporting user agent: the user agent must add all the style
    // sheets that apply to the iframe element to the cascade of the active document
    // of the iframe element's nested browsing context, at the appropriate cascade
    // levels, before any style sheets specified by the document itself.
    shouldBeEqualToString("window.getComputedStyle(one).color", "rgb(255, 0, 0)"); // red, Specified directly by parent's selector.
    shouldBeEqualToString("window.getComputedStyle(two).color", "rgb(128, 0, 128)"); // purple, Selector in child overrides parent.
    shouldBeEqualToString("window.getComputedStyle(three).color", "rgb(255, 255, 255)"); // white, div selector in parent.

    // Spec: In a CSS-supporting user agent: the user agent must, for the purpose of
    // CSS property inheritance only, treat the root element of the active document
    // of the iframe element's nested browsing context as being a child of the
    // iframe element. (Thus inherited properties on the root element of the
    // document in the iframe will inherit the computed values of those properties
    // on the iframe element instead of taking their initial values.)
    window.rootElement = iframe.contentDocument.documentElement;
    shouldBeEqualToString("window.getComputedStyle(rootElement).color", "rgb(255, 128, 64)"); // orange, inherited from parent iframe.
}
</script>
